[UPDATE] sceneSize method return double -> unsigned int

+++ include/IScene.hpp
@@ -22,8 +22,8 @@ namespace arcade
         virtual ~IScene() = default;
 
         virtual std::vector<std::unique_ptr<arcade::IEntity>> &getEntities() = 0;
-        virtual const double &getSceneWidth() = 0;
-        virtual const double &getSceneHeight() = 0;
+        virtual const uint32_t &getSceneWidth() = 0;
+        virtual const uint32_t &getSceneHeight() = 0;
 
     protected:
     private:
